Fix weird image output (#15)
Use f64 instead of truncating to integers everywhere

* Add test fixes

* - Use f64 consistently until the very moment we draw a pixel.
- Add convenience function for faster PNG writing
- Update gifski to more recent version to unbreak koch_curve_video
  example

* No reason to specify this type here

* Minor changes

+++ dcc-lsystem/src/renderer.rs
@@ -75,7 +75,7 @@ impl<Q: TurtleContainer> TurtleRenderer<Q> {
 pub struct DataRendererOptions {}
 
 impl<Q: TurtleContainer> Renderer<DataRendererOptions> for TurtleRenderer<Q> {
-    type Output = Vec<(i32, i32, i32, i32)>;
+    type Output = Vec<(f64, f64, f64, f64)>;
 
     fn render(mut self, system: &LSystem, _options: &DataRendererOptions) -> Self::Output {
         // Setup our state machine based on the LSystem state
